@@ -1,12 +1,12 @@
 module github.com/gochain/gochain/v4
 
-go 1.20
+go 1.22.7
 
 require (
-	contrib.go.opencensus.io/exporter/stackdriver v0.6.0
+	contrib.go.opencensus.io/exporter/stackdriver v0.13.14
 	github.com/allegro/bigcache v1.2.1
 	github.com/aristanetworks/goarista v0.0.0-20180424004133-70dca2f27708
-	github.com/aws/aws-sdk-go v1.34.0 // indirect
+	github.com/aws/aws-sdk-go v1.43.31 // indirect
 	github.com/cespare/cp v1.0.0
 	github.com/cespare/xxhash v1.1.0
 	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
